refactor: split code in different modules
