+++ src/user/user.controller.ts
@@ -1,0 +1,18 @@
+import { Controller, Get, Param } from '@nestjs/common';
+import { UserService } from './user.service';
+import { User } from './user.interface';
+
+@Controller('user')
+export class UserController {
+  constructor(private readonly userService: UserService) {}
+
+  @Get()
+  getUsers(): User[] {
+    return this.userService.getUsers();
+  }
+
+  @Get(':id')
+  getUser(@Param('id') id: string) {
+    return this.userService.getUser(id);
+  }
+}
